requirements.txt
customtkinter
Pillow
pyzbar
qrcode
